requirements.txt
asgiref==3.5.2
Django==4.0.8
sqlparse==0.4.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
django-allauth==0.50.0
dj-rest-auth==2.2.5
django-crispy-forms==1.14.0
crispy_bootstrap5==0.7
psycopg2-binary==2.9.3
environs[django]==9.5.0
pillow==9.0.1
django-debug-toolbar==3.4.0
drf-spectacular==0.24.2
whitenoise==6.1.0
gunicorn==20.1.0
django-reversion==5.0.3